# services/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aantoonbaar API",
    "description": "Python-backend voor de Aantoonbaar eisen-bewijsmatrix.",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/projects": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Project Index",
        "operationId": "project_index_api_v1_projects_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectResponse"
                  },
                  "title": "Response Project Index Api V1 Projects Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Project Create",
        "operationId": "project_create_api_v1_projects_post",
        "parameters": [
          {
            "name": "idempotency-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Project Detail",
        "operationId": "project_detail_api_v1_projects__project_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/workspace": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Project Workspace",
        "operationId": "project_workspace_api_v1_projects__project_id__workspace_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/source-documents": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Source Document Create",
        "operationId": "source_document_create_api_v1_projects__project_id__source_documents_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_source_document_create_api_v1_projects__project_id__source_documents_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/source-documents/{document_id}/extract": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Source Document Extract",
        "operationId": "source_document_extract_api_v1_projects__project_id__source_documents__document_id__extract_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MutationResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/requirements": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Requirement Create",
        "operationId": "requirement_create_api_v1_projects__project_id__requirements_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequirementCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MutationResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/requirements/{requirement_id}": {
      "patch": {
        "tags": [
          "projects"
        ],
        "summary": "Requirement Update",
        "operationId": "requirement_update_api_v1_projects__project_id__requirements__requirement_id__patch",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "requirement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Requirement Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequirementUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MutationResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "projects"
        ],
        "summary": "Requirement Delete",
        "operationId": "requirement_delete_api_v1_projects__project_id__requirements__requirement_id__delete",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "requirement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Requirement Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MutationResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/requirements/merge": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Requirement Merge",
        "operationId": "requirement_merge_api_v1_projects__project_id__requirements_merge_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequirementsMerge"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MutationResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/evidence-documents": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Evidence Document Create",
        "operationId": "evidence_document_create_api_v1_projects__project_id__evidence_documents_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_evidence_document_create_api_v1_projects__project_id__evidence_documents_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceDocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/evidence-documents/{document_id}/file": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Evidence Document File",
        "operationId": "evidence_document_file_api_v1_projects__project_id__evidence_documents__document_id__file_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/matches/generate": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Matches Generate",
        "operationId": "matches_generate_api_v1_projects__project_id__matches_generate_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MutationResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/requirements/{requirement_id}/assessment": {
      "put": {
        "tags": [
          "projects"
        ],
        "summary": "Assessment Save",
        "operationId": "assessment_save_api_v1_projects__project_id__requirements__requirement_id__assessment_put",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "requirement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Requirement Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssessmentUpsert"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MutationResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/export.csv": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Project Export",
        "operationId": "project_export_api_v1_projects__project_id__export_csv_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProjectCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "client": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Client"
          },
          "reference": {
            "type": "string",
            "maxLength": 100,
            "title": "Reference",
            "default": ""
          },
          "dueDate": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duedate"
          },
          "product": {
            "type": "string",
            "maxLength": 200,
            "title": "Product",
            "default": ""
          },
          "productVersion": {
            "type": "string",
            "maxLength": 100,
            "title": "Productversion",
            "default": ""
          },
          "owner": {
            "type": "string",
            "maxLength": 200,
            "title": "Owner",
            "default": ""
          },
          "notes": {
            "type": "string",
            "maxLength": 5000,
            "title": "Notes",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "name",
          "client"
        ],
        "title": "ProjectCreate"
      },
      "ProjectResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "client": {
            "type": "string",
            "title": "Client"
          },
          "reference": {
            "type": "string",
            "title": "Reference"
          },
          "dueDate": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duedate"
          },
          "product": {
            "type": "string",
            "title": "Product"
          },
          "productVersion": {
            "type": "string",
            "title": "Productversion"
          },
          "owner": {
            "type": "string",
            "title": "Owner"
          },
          "notes": {
            "type": "string",
            "title": "Notes"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "requirementCount": {
            "type": "integer",
            "title": "Requirementcount",
            "default": 0
          },
          "reviewedCount": {
            "type": "integer",
            "title": "Reviewedcount",
            "default": 0
          },
          "sufficientCount": {
            "type": "integer",
            "title": "Sufficientcount",
            "default": 0
          },
          "missingCount": {
            "type": "integer",
            "title": "Missingcount",
            "default": 0
          },
          "nextDeadline": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextdeadline"
          },
          "progress": {
            "type": "integer",
            "title": "Progress",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "client",
          "reference",
          "dueDate",
          "product",
          "productVersion",
          "owner",
          "notes",
          "createdAt"
        ],
        "title": "ProjectResponse"
      },
...
      "WorkspaceResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "client": {
            "type": "string",
            "title": "Client"
          },
          "reference": {
            "type": "string",
            "title": "Reference"
          },
          "dueDate": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duedate"
          },
          "product": {
            "type": "string",
            "title": "Product"
          },
          "productVersion": {
            "type": "string",
            "title": "Productversion"
          },
          "owner": {
            "type": "string",
            "title": "Owner"
          },
          "notes": {
            "type": "string",
            "title": "Notes"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "requirementCount": {
            "type": "integer",
            "title": "Requirementcount",
            "default": 0
          },
          "reviewedCount": {
            "type": "integer",
            "title": "Reviewedcount",
            "default": 0
          },
          "sufficientCount": {
            "type": "integer",
            "title": "Sufficientcount",
            "default": 0
          },
          "missingCount": {
            "type": "integer",
            "title": "Missingcount",
            "default": 0
          },
          "nextDeadline": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextdeadline"
          },
          "progress": {
            "type": "integer",
            "title": "Progress",
            "default": 0
          },
          "sourceDocuments": {
            "items": {
              "$ref": "#/components/schemas/SourceDocumentResponse"
            },
            "type": "array",
            "title": "Sourcedocuments"
          },
          "evidenceDocuments": {
            "items": {
              "$ref": "#/components/schemas/EvidenceDocumentResponse"
            },
            "type": "array",
            "title": "Evidencedocuments"
          },
          "requirements": {
            "items": {
              "$ref": "#/components/schemas/RequirementResponse"
            },
            "type": "array",
            "title": "Requirements"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/TaskResponse"
            },
            "type": "array",
            "title": "Tasks"
          },
          "auditEvents": {
            "items": {
              "$ref": "#/components/schemas/AuditEventResponse"
            },
            "type": "array",
            "title": "Auditevents"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "client",
          "reference",
          "dueDate",
          "product",
          "productVersion",
          "owner",
          "notes",
          "createdAt",
          "sourceDocuments",
          "evidenceDocuments",
          "requirements",
          "tasks",
          "auditEvents"
        ],
        "title": "WorkspaceResponse"
      }
    }
  }
}